requirements.txt
yt-dlp>=2023.11.16
opencv-python>=4.8.1
ultralytics>=8.0.0
openai>=1.3.0
streamlit>=1.28.0
plotly>=5.17.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
